fix(build): prevent pre-publish TS error
When trying to run `npm publish`, received the following TS error:

```
TS2322: Type 'ReactSelectProps<unknown>' is not assignable to type 'ReactSelectProps<OptionValues>'.
```

Inline the call to `onInputChange` in both tests so we don't need to assign the props a type. Verified with `npm publish --dry-run` that this resolved the TS error.